removed support for aks as it is deprecated
removed support for aks as it is deprecated
