--- examples/service/ocean/providers/azure/read/main.go
@@ -1,45 +1,0 @@
-package main
-
-import (
-	"context"
-	"log"
-
-	"github.com/spotinst/spotinst-sdk-go/service/ocean"
-	"github.com/spotinst/spotinst-sdk-go/service/ocean/providers/azure"
-	"github.com/spotinst/spotinst-sdk-go/spotinst"
-	"github.com/spotinst/spotinst-sdk-go/spotinst/session"
-	"github.com/spotinst/spotinst-sdk-go/spotinst/util/stringutil"
-)
-
-func main() {
-	// All clients require a Session. The Session provides the client with
-	// shared configuration such as account and credentials.
-	// A Session should be shared where possible to take advantage of
-	// configuration and credential caching. See the session package for
-	// more information.
-	sess := session.New()
-
-	// Create a new instance of the service's client with a Session.
-	// Optional spotinst.Config values can also be provided as variadic
-	// arguments to the New function. This option allows you to provide
-	// service specific configuration.
-	svc := ocean.New(sess)
-
-	// Create a new context.
-	ctx := context.Background()
-
-	// Read cluster configuration.
-	out, err := svc.CloudProviderAzure().ReadCluster(ctx, &azure.ReadClusterInput{
-		ClusterID: spotinst.String("o-12345"),
-	})
-	if err != nil {
-		log.Fatalf("spotinst: failed to read cluster: %v", err)
-	}
-
-	// Output.
-	if out.Cluster != nil {
-		log.Printf("Cluster %q: %s",
-			spotinst.StringValue(out.Cluster.ID),
-			stringutil.Stringify(out.Cluster))
-	}
-}
