+++ examples/service/ocean/providers/azure_np/stoproll/main.go
@@ -2,13 +2,13 @@ package main
 
 import (
 	"context"
+	"github.com/spotinst/spotinst-sdk-go/service/ocean/providers/azure_np"
+	"github.com/spotinst/spotinst-sdk-go/spotinst/util/stringutil"
 	"log"
 
 	"github.com/spotinst/spotinst-sdk-go/service/ocean"
-	"github.com/spotinst/spotinst-sdk-go/service/ocean/providers/azure"
 	"github.com/spotinst/spotinst-sdk-go/spotinst"
 	"github.com/spotinst/spotinst-sdk-go/spotinst/session"
-	"github.com/spotinst/spotinst-sdk-go/spotinst/util/stringutil"
 )
 
 func main() {
@@ -28,26 +28,20 @@ func main() {
 	// Create a new context.
 	ctx := context.Background()
 
-	// List right sizing suggestions.
-	out, err := svc.CloudProviderAzure().ListResourceSuggestions(ctx, &azure.ListResourceSuggestionsInput{
-		OceanID: spotinst.String("o-12345"),
-		Filter: &azure.Filter{
-			Attribute: &azure.Attribute{
-				Key:      spotinst.String("foo"),
-				Operator: spotinst.String("foo"),
-				Type:     spotinst.String("foo"),
-				Value:    spotinst.String("foo"),
-			},
-			Namespaces: []string{"foo"},
-		},
+	// Stop cluster roll.
+	out, err := svc.CloudProviderAzureNP().StopRoll(ctx, &azure_np.StopRollInput{
+		ClusterID: spotinst.String("o-12345"),
+		RollID:    spotinst.String("scr-7890"),
 	})
 	if err != nil {
-		log.Fatalf("spotinst: failed to list right sizing suggestions: %v", err)
+		log.Fatalf("spotinst: failed to stop roll: %v", err)
 	}
 
-	// Output.
-	if out.Suggestions != nil {
-		log.Printf("Right Sizing Suggestions: %s",
-			stringutil.Stringify(out.Suggestions))
+	if len(out.Rolls) > 0 {
+		for _, roll := range out.Rolls {
+			log.Printf("Roll %q: %s",
+				spotinst.StringValue(roll.ID),
+				stringutil.Stringify(roll))
+		}
 	}
 }
