+++ service/ocean/ocean.go
@@ -2,7 +2,6 @@ package ocean
 
 import (
 	"github.com/spotinst/spotinst-sdk-go/service/ocean/providers/aws"
-	"github.com/spotinst/spotinst-sdk-go/service/ocean/providers/azure"
 	"github.com/spotinst/spotinst-sdk-go/service/ocean/providers/azure_np"
 	"github.com/spotinst/spotinst-sdk-go/service/ocean/providers/gcp"
 	"github.com/spotinst/spotinst-sdk-go/service/ocean/spark"
@@ -17,7 +16,6 @@ import (
 type Service interface {
 	CloudProviderAWS() aws.Service
 	CloudProviderGCP() gcp.Service
-	CloudProviderAzure() azure.Service
 	Spark() spark.Service
 	CloudProviderAzureNP() azure_np.Service
 }
@@ -46,12 +44,6 @@ func (s *ServiceOp) CloudProviderAWS() aws.Service {
 
 func (s *ServiceOp) CloudProviderGCP() gcp.Service {
 	return &gcp.ServiceOp{
-		Client: s.Client,
-	}
-}
-
-func (s *ServiceOp) CloudProviderAzure() azure.Service {
-	return &azure.ServiceOp{
 		Client: s.Client,
 	}
 }
