+++ service/ocean/providers/azure_np/service.go
@@ -17,6 +17,10 @@ type Service interface {
 	UpdateCluster(context.Context, *UpdateClusterInput) (*UpdateClusterOutput, error)
 	DeleteCluster(context.Context, *DeleteClusterInput) (*DeleteClusterOutput, error)
 	ImportCluster(context.Context, *ImportClusterInput) (*ImportClusterOutput, error)
+	CreateRoll(context.Context, *CreateRollInput) (*CreateRollOutput, error)
+	ReadRoll(ctx context.Context, input *ReadRollInput) (*ReadRollOutput, error)
+	ListRolls(ctx context.Context, input *ListRollsInput) (*ListRollsOutput, error)
+	StopRoll(ctx context.Context, input *StopRollInput) (*StopRollOutput, error)
 
 	ListVirtualNodeGroups(context.Context, *ListVirtualNodeGroupsInput) (*ListVirtualNodeGroupsOutput, error)
 	CreateVirtualNodeGroup(context.Context, *CreateVirtualNodeGroupInput) (*CreateVirtualNodeGroupOutput, error)
